% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R, R/rtoqmd_addin.R
\name{rtoqmd_addin}
\alias{rtoqmd_addin}
\title{Convert Active R Script to Quarto Markdown}
\usage{
rtoqmd_addin()

rtoqmd_addin()
}
\description{
RStudio add-in that converts the currently active R script in the editor
to a Quarto markdown document. Uses a Shiny interface for parameter input.

RStudio add-in that converts the currently active R script in the editor
to a Quarto markdown document. Uses a Shiny interface for parameter input.
}
